ZetaSspBidAdapter: rename files across ${bidderCode}BidAdapter.js (#7317)
